backend/requirements.txt
asgiref==3.8.1
Django==5.2.7
django-cors-headers==4.6.0
djangorestframework==3.15.2
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
python-dotenv==1.0.1
django-phonenumber-field==8.3.0 
phonenumbers==9.0.16